Add URI Validator Tests (#19)
* Added URI Validator Tests

* Added URI Validator (UriValidator) Tests
* Minor Support to MacOS

* Remove failing Lint Exception

* Add Extra URIValidator Cases, 1.5.7 Update

* Add URIValidator cases to check is_micro_form, is_long_form, etc
* Update Java/Python Test Agents to 1.5.7

The additional URIValidator test cases are necessary since they provide more coverage of URIValidator methods. The update to 1.5.7 was necessary to align with the latest releases.

* Removed unnecessary if/else statement

Thanks to Neelam's advice, I noticed that I was doing an unnecessary step. Now, was able to simplify by removing if/else statement in the test agents.

---------

Co-authored-by: Matthew D'Alonzo <matthew.dalonzo@gm.com>