Implement SocketUTransport for mock testing. (#64)
This PR implements a socket transport to the python dispacther script,
and necessary data structures for listener callbacks and callback
removal. The high level code is a private implementation in
SocketUTransport.cpp. The unordered_map supporting callbacks is in
SafeTupleMap.h. The details of this is std::hash extensions to hash
tuples of optionals for filtering with wildcards.